+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,1 @@
-*   Remove the need to use reflection within ProgressPrinter for appending Sub-Modules to a parent Module
+*   Throw an exception when defining duplicate constraint keys on the same NotInParallel attribute